add signal file
