+++ rental_site/main/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class MainConfig(AppConfig):
     name = 'main'
+
+    def ready(self):
+        from . import signals
