+++ rental_site/main/views.py
@@ -7,9 +7,6 @@ from django.forms.models import inlineformset_factory
 from django.shortcuts import render, redirect
 from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
 from django.contrib import messages
-from django.dispatch import receiver
-from django.db.models.signals import post_save
-from django.contrib.auth.models import User, Group
 
 from .models import *
 from .forms import *
@@ -302,16 +299,3 @@ class EditRealtyAd(LoginRequiredMixin, PermissionRequiredMixin, UpdateView):
             messages.error(request, f'Ошибка! {ad_form.errors.as_text()}')
 
         return self.render_to_response(self.get_context_data())
-
-
-@receiver(post_save, sender=User)
-def create_profile(sender, instance, created, **kwargs):
-    if created:
-        user_profile = Profile.objects.create(user=instance)
-        user_profile.save()
-
-
-@receiver(post_save, sender=User)
-def add_user_to_group(sender, instance, created, **kwargs):
-    common_users, _ = Group.objects.get_or_create(name="common_users")
-    instance.groups.add(common_users)
